Move RequestDecorator to common module
