+++ features/letterboxd/api/build.gradle
@@ -40,6 +40,7 @@ dependencies {
     implementation G.maven.retrofit.moshi
     implementation G.maven.retrofit.rxjava
     implementation G.maven.rxjava
+    implementation project(':features:common:api')
 
     testImplementation G.maven.junit
     testImplementation G.maven.truth
