+++ features/letterboxd/api/src/main/kotlin/com/letterboxd/authentication/UrlRequestDecorator.kt
@@ -1,5 +1,6 @@
 package com.letterboxd.authentication
 
+import io.archano.playground.common.api.RequestDecorator
 import okhttp3.Request
 import java.util.*
 import java.util.concurrent.TimeUnit
